rsx: Minor improvements to texture cache
- Adds depth to temp subresource key to avoid 3D mismatch (resample)
- Fix narrowing warning for block_h calculations

+++ rpcs3/Emu/RSX/Common/texture_cache.h
@@ -2090,7 +2090,7 @@ namespace rsx
 			// Check surface cache early if the option is enabled
 			if (options.prefer_surface_cache)
 			{
-				const u16 block_h = (attr.depth * attr.slice_h);
+				const u32 block_h = (attr.depth * attr.slice_h);
 				overlapping_fbos = m_rtts.get_merged_texture_memory_region(cmd, attr.address, attr.width, block_h, attr.pitch, attr.bpp, rsx::surface_access::shader_read);
 
 				if (!overlapping_fbos.empty())
@@ -2151,7 +2151,7 @@ namespace rsx
 			if (!options.prefer_surface_cache)
 			{
 				// Now check for surface cache hits
-				const u16 block_h = (attr.depth * attr.slice_h);
+				const u32 block_h = (attr.depth * attr.slice_h);
 				overlapping_fbos = m_rtts.get_merged_texture_memory_region(cmd, attr.address, attr.width, block_h, attr.pitch, attr.bpp, rsx::surface_access::shader_read);
 			}
